feat(ignition): robust launch module — atomic tier commit, idempotent launch, emergency drain
Implements 4 top-ranked robustness ideas (avg score 93/100):
- Idea #16 (Score 97): Atomic Tier Commit Protocol
- Idea #21 (Score 93): Idempotent Launch with State Reconciliation
- Idea #29 (Score 88): Post-Launch Stabilization Window
- Idea #30 (Score 94): Launch Abort with Emergency Drain

robust_launch.rs (650+ lines):
- launch_container_idempotent(): skip if running, reconcile if wrong state
- launch_tier_atomic(): if ANY container fails, rollback ALL started containers
- stabilization_window(): continuous health monitoring after all containers running
- emergency_drain(): stop all containers in REVERSE tier order, clean networks, preserve volumes
- Bounded concurrency via semaphore (MAX_CONCURRENT_LAUNCHES = 4)
- 9 unit tests

types.rs expanded:
- CascadeState, PartitionResult, DrainResult, IgnitionCheckpoint
- KnownGoodConfig, ContainerStateSnapshot, TierCommitResult
- Constants: MAX_CASCADE_DEPTH, EMERGENCY_DRAIN_TIMEOUT_MS,
  STABILIZATION_WINDOW_MS, MAX_CONCURRENT_LAUNCHES

podman.rs expanded:
- container_image, container_id, remove_container, run_container
- list_networks, remove_network, force_remove

cascade.rs (497 lines) — Cascading failure containment (Idea #46, Score 98)
connectivity.rs (497 lines) — Inter-container connectivity matrix (Idea #32, Score 92)
  - Zenoh mesh topology verification (Idea #34, Score 93)

All 240 tests passing, 0 compilation errors.

+++ sub-projects/intelitor-v5.2/native/ignition_daemon/src/connectivity.rs
@@ -254,7 +254,7 @@ pub async fn verify_zenoh_mesh_topology() -> Result<ZenohMeshReport, IgnitionErr
             report.peer_visibility.push(ZenohPeerVisibility {
                 router: router.to_string(),
                 visible_peers: 0,
-                expected_peers: routers.len() - 1,
+                expected_peers: (routers.len() - 1) as u32,
                 healthy: false,
             });
             continue;
